refactor: move render abstraction to multik::render
